Plot graph in summary

+++ Assets/Scripts/Game/GameModeBase.cs
@@ -161,7 +161,7 @@ public class GameModeBase : MonoBehaviour
             }
         );
 
-        double progress = (this._player.Resources.Food * 1) + 
+        float progress = (this._player.Resources.Food * 1) + 
                         (this._player.Resources.Gold * 1.5f) + 
                         (this._player.Resources.Military * 2) + 
                         (this._player.Resources.People * 3) +
